Reorganise the structure and get it to compile cleanly
