+++ src/externs/externs.js
@@ -1,0 +1,4 @@
+// Allow use of module.exports for Node.
+var module = {
+    exports: {}
+};
