+++ src/rendering-function.js
@@ -1,0 +1,14 @@
+goog.provide('DubStash.ExternalRenderingFunction');
+goog.provide('DubStash.ContextualRenderingFunction');
+
+
+/**
+ * @typedef {function(Object, boolean=, DubStash.runtime.Context=):string}
+ */
+DubStash.ExternalRenderingFunction;
+
+
+/**
+ * @typedef {function(DubStash.runtime.Context, boolean=):string}
+ */
+DubStash.ContextualRenderingFunction;
